feat: streamline existing permissions handling and update confirmation dialog logic
