+++ packages/gator-permissions-snap/src/core/existingpermissions/index.ts
@@ -1,3 +1,4 @@
+export { ExistingPermissionsState } from './existingPermissionsState';
 export { ExistingPermissionsService } from './existingPermissionsService';
 export {
   buildExistingPermissionsContent,
